Include ROWSKIPS and ROWCOUNT into RSAPReadTable
In order to skip rows and limit the number of rows to be retrieved, one can use the options ROWSKIPS and ROWCOUNT of RFC_READ_TABLE.
One has to bear in mind though that depending on the NW RFC version there could be limits on these options based on their types - I've read of instances where ROWSKIPS is INT4 and thus limited to 999999. The version that is available to me does not seem to impose such restrictions.